Create antgame.py
You only need to change the code at line 41-42 to give your ant a food gathering strategy  In the real world, this would be coded in genes.  In our world, we do it in Python.  This program also watches the ant and saves a few files with statistics.  More about that later.